test: `partial` witnessed from JavaScript; the duplicate guard coerces like the marshal; liblloyal → cee612d
The last hop of the rc/partial contract, asserted on real weights from the
JS side. test/integration.ts gains testDecodeFailure(): a cohort whose
third scatter chunk finds no KV slot rejects with rc 1 and partial true —
the landed children moved and were charged, the refused one did not and
still prefills, prune returns the pool to zero; a 300-token prefill that
dies after four 64-token chunks reports partial with its books unmoved,
and a fresh branch prefills after prune. Both run on the default model
and, when present, on Qwen3.5-4B (the production default, a Gated
DeltaNet hybrid). The media rail is asserted per entry: an image costing
576 cells against a 256-cell pool comes back with error, rc 1 and
partial true, and prune reclaims the branch. nSeqMax stays at 4 or 2.

_storePrefillMultimodal's duplicate-handle guard compared DoubleValue()
while the marshal coerced with Uint32Value(), so 5 and 5.5 passed the
guard and named one branch. The guard coerces the same way now; the
integration test asserts the fractional duplicate is rejected.

tsconfig.test.json emitted JavaScript beside the sources on every
build:test — outDir "." pulled the two src/ modules the unit test imports
into the emit. The gate is noEmit: nothing consumed the output (the
suites run through tsx). loadBinary's `variant` doc matches GpuVariant.

Submodule pointer: cee612d (create refuses a no-vocab model; rc -1
restores too; the real-weights failure suite). Addon rebuilt against it.

+++ src/SessionContext.cpp
@@ -726,7 +726,8 @@ public:
      *  before its decode ran (bitmap decode, marker/bitmap mismatch: the
      *  branch is untouched) or inside decode_segments. `rc` and `partial`
      *  say which case a decode failure is (see DecodeError in liblloyal):
-     *  intact iff rc == 1 && !partial; anything else ⇒ prune and replay from
+     *  intact iff the failing call restored state (rc 1 or -1) and nothing
+     *  before it landed (!partial); anything else ⇒ prune and replay from
      *  content — pruning an untouched branch is safe. */
     std::string error;
     /** llama_decode's raw return code when the failure carried one (0
@@ -2776,10 +2777,12 @@ Napi::Value SessionContext::_storePrefillMultimodal(const Napi::CallbackInfo& in
   // dispatches one handle at a time so the pair never meets there — check it
   // here to keep the same fail-loud contract as _storePrefill/_storeCommit.
   {
-    std::vector<double> seen;
+    // Coerced exactly as the marshal below coerces them: 5 and 5.5 name the
+    // same branch, and the guard must see that.
+    std::vector<uint32_t> seen;
     seen.reserve(n);
     for (uint32_t i = 0; i < n; i++) {
-      const double h = jsHandles.Get(i).As<Napi::Number>().DoubleValue();
+      const uint32_t h = jsHandles.Get(i).As<Napi::Number>().Uint32Value();
       for (uint32_t j = 0; j < seen.size(); j++) {
         if (seen[j] == h) {
           throw Napi::Error::New(env,
